Added latest changes requested by Jeff, so directory renamed and no global API
