--- session_manager/go.mod
@@ -1,5 +1,0 @@
-module gotests/amass-engine/session_manager
-
-go 1.21.1
-
-require github.com/google/uuid v1.3.1 // indirect
