+++ sessions/session_manager_test.go
@@ -1,0 +1,29 @@
+package sessions
+
+import (
+	"fmt"
+	"testing"
+)
+
+var (
+	ss *SessionStorage
+)
+
+func TestMain(m *testing.M) {
+	ss := NewSessionStorage()
+	defer ss.Shutdown()
+
+	m.Run()
+}
+
+func TestAddSession001(t *testing.T) {
+	// Create a session
+	s := &Config{
+		// ...
+	}
+	id := ss.AddSession(s)
+	fmt.Println("Session ID:", id)
+	if id == zeroSessionUUID {
+		t.Error("Session ID is zero")
+	}
+}
